Update PR template for Riff; update ci workflow, add release workflow
2022-08-30 increased -max-old-space-size to 5120 because it was failing again
2022-04-20 setting NODE_OPTIONS --max-old-space-size=4096 because the ci action was failing with:
```
FATAL ERROR: Reached heap limit Allocation failed - JavaScript heap out of memory
```

This is a max limit that causes the garbage collector to kick in. Set it so it leaves enough memory for other things (the github runner has ~7GB) but not so small that there isn't enough room for the things that are needed. I tried 4096 first and it worked well, but I'd guess you could kick it up to 5.5 or 6GB if you wanted.