build: update release github action
release update addresses this issue:
```
The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/
```
by fixing the action's use of set-output, and by replacing the actions/create-release and actions/upload-release-asset actions with the ncipollo/release-action action found in the marketplace. I've looked at the repo and it is open source AND looks well maintained at this time.